Added minor fixes
